% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hindi_morph.R
\name{hindi_morph}
\alias{hindi_morph}
\title{Morphological Analysis Hindi Using RDRPOST}
\usage{
hindi_morph(data, text_col)
}
\arguments{
\item{data}{dataframe contains a column to analyze}

\item{text_col}{column name of the text toanalyze}
}
\description{
Morphological Analysis Hindi Using RDRPOST
}
\author{
Yuya Matsumura
}
